refactor: orderedmap package -> omap
