+++ internal/omap/orderedmap_test.go
@@ -1,4 +1,4 @@
-package orderedmap
+package omap
 
 import (
 	"testing"
